requirements.txt
langchain==0.3.0
langchain-community
langchain-openai
openai
tiktoken

pypdf
pdf2image
pymupdf
docx2txt
openpyxl
beautifulsoup4
requests

faiss-cpu

python-dotenv
streamlit
Pillow
matplotlib

sentence-transformers
